main/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
cohere==4.11.0
pdfplumber==0.10.2
PyPDF2==2.11.2
transformers==4.34.0
torch==2.1.0
scikit-learn==1.3.1
numpy==1.24.4
gunicorn==21.2.0
